Modularize TreeView and TreeViewDU
