+++ packages/ssz/src/index.ts
@@ -22,7 +22,9 @@ export {ByteArrayType} from "./type/byteArray";
 // Base type clases
 export {Type, ValueOf, JsonPath} from "./type/abstract";
 export {BasicType, isBasicType} from "./type/basic";
-export {CompositeType, TreeView, TreeViewDU, CompositeView, CompositeViewDU, isCompositeType} from "./type/composite";
+export {CompositeType, CompositeView, CompositeViewDU, isCompositeType} from "./type/composite";
+export {TreeView} from "./view/abstract";
+export {TreeViewDU} from "./viewDU/abstract";
 
 // Values
 export {BitArray, getUint8ByteToBitBooleanArray} from "./value/bitArray";
