+++ packages/ssz/src/type/composite.ts
@@ -11,7 +11,7 @@ import {
 import {byteArrayEquals} from "../util/byteArray";
 import {merkleize} from "../util/merkleize";
 import {treePostProcessFromProofNode} from "../util/proof/treePostProcessFromProofNode";
-import {Type, ValueOf, ByteViews, JsonPath, JsonPathProp} from "./abstract";
+import {Type, ByteViews, JsonPath, JsonPathProp} from "./abstract";
 export {ByteViews};
 
 export const LENGTH_GINDEX = BigInt(3);
@@ -37,8 +37,6 @@ export type CompositeViewDU<T extends CompositeType<unknown, unknown, unknown>> 
 /** Any CompositeType without any generic arguments */
 export type CompositeTypeAny = CompositeType<unknown, unknown, unknown>;
 
-/* eslint-disable @typescript-eslint/member-ordering  */
-
 /** Dedicated property to cache hashTreeRoot of immutable CompositeType values */
 const symbolCachedPermanentRoot = Symbol("ssz_cached_permanent_root");
 
@@ -46,6 +44,8 @@ const symbolCachedPermanentRoot = Symbol("ssz_cached_permanent_root");
 type ValueWithCachedPermanentRoot = {
   [symbolCachedPermanentRoot]?: Uint8Array;
 };
+
+/* eslint-disable @typescript-eslint/member-ordering  */
 
 /**
  * Represents a composite type as defined in the spec:
@@ -344,141 +344,6 @@ export abstract class CompositeType<V, TV, TVDU> extends Type<V> {
   abstract getPropertyType(property: JsonPathProp): Type<unknown>;
   /** Return a leaf node index's property if the index is within bounds */
   abstract getIndexProperty(index: number): JsonPathProp | null;
-}
-
-/**
- * A Tree View is a wrapper around a type and an SSZ Tree that contains:
- * - data merkleized
- * - a hook to its parent Tree to propagate changes upwards
- *
- * **View**
- * - Best for simple usage where performance is NOT important
- * - Applies changes immediately
- * - Has reference to parent tree
- * - Does NOT have caches for fast get / set ops
- */
-export abstract class TreeView<T extends CompositeType<unknown, unknown, unknown>> {
-  /** Merkle tree root node */
-  abstract readonly node: Node;
-  /** SSZ type associated with this Tree View */
-  abstract readonly type: T;
-
-  /** Serialize view to binary data */
-  serialize(): Uint8Array {
-    const output = new Uint8Array(this.type.tree_serializedSize(this.node));
-    const dataView = new DataView(output.buffer, output.byteOffset, output.byteLength);
-    this.type.tree_serializeToBytes({uint8Array: output, dataView}, 0, this.node);
-    return output;
-  }
-
-  /**
-   * Merkleize view and compute its hashTreeRoot.
-   *
-   * See spec for definition of hashTreeRoot:
-   * https://github.com/ethereum/consensus-specs/blob/dev/ssz/simple-serialize.md#merkleization
-   */
-  hashTreeRoot(): Uint8Array {
-    return this.node.root;
-  }
-
-  /**
-   * Create a Merkle multiproof on this view's data.
-   * A `path` is an array of 'JSON' paths into the data
-   * @example
-   * ```ts
-   * state.createProof([
-   *   ["validators", 1234, "slashed"],
-   *   ["genesisTime"]
-   * ])
-   * ```
-   *
-   * See spec for definition of merkle multiproofs:
-   * https://github.com/ethereum/consensus-specs/blob/dev/ssz/merkle-proofs.md#merkle-multiproofs
-   */
-  createProof(paths: JsonPath[]): Proof {
-    return this.type.tree_createProof(this.node, paths);
-  }
-
-  /**
-   * Transform the view into a value, from the current node instance.
-   * For ViewDU returns the value of the committed data, so call .commit() before if there are pending changes.
-   */
-  toValue(): ValueOf<T> {
-    return this.type.tree_toValue(this.node) as ValueOf<T>;
-  }
-
-  /** Return a new Tree View instance referencing the same internal `Node`. Drops its existing `Tree` hook if any */
-  clone(): this {
-    return this.type.getView(new Tree(this.node)) as this;
-  }
-}
-
-/**
- * A Deferred Update Tree View (`ViewDU`) is a wrapper around a type and
- * a SSZ Node that contains:
- * - data merkleized
- * - some arbitrary caches to speed up data manipulation required by the type
- *
- * **ViewDU**
- * - Best for complex usage where performance is important
- * - Defers changes to when commit is called
- * - Does NOT have a reference to the parent ViewDU
- * - Has caches for fast get / set ops
- */
-export abstract class TreeViewDU<T extends CompositeType<unknown, unknown, unknown>> extends TreeView<T> {
-  /**
-   * Applies any deferred updates that may be pending in this ViewDU instance and updates its internal `Node`.
-   */
-  abstract commit(): void;
-
-  /**
-   * Returns arbitrary data that is useful for this ViewDU instance to optimize data manipulation. This caches MUST
-   * not include non-commited data. `this.cache` can be called at any time, both before and after calling `commit()`.
-   */
-  abstract readonly cache: unknown;
-
-  /**
-   * MUST drop any reference to mutable cache data. After `clearCache()`, if the dropped caches are mutated, no changes
-   * should apply to this instance both before and after calling `commit()`.
-   */
-  protected abstract clearCache(): void;
-
-  /**
-   * Merkleize view and compute its hashTreeRoot.
-   * Commits any pending changes before computing the root.
-   *
-   * See spec for definition of hashTreeRoot:
-   * https://github.com/ethereum/consensus-specs/blob/dev/ssz/simple-serialize.md#merkleization
-   */
-  hashTreeRoot(): Uint8Array {
-    this.commit();
-    return super.hashTreeRoot();
-  }
-
-  /**
-   * Serialize view to binary data.
-   * Commits any pending changes before computing the root.
-   */
-  serialize(): Uint8Array {
-    this.commit();
-    return super.serialize();
-  }
-
-  /**
-   * Return a new ViewDU instance referencing the same internal `Node`.
-   *
-   * By default it will transfer the cache of this ViewDU to the new cloned instance. Set `dontTransferCache` to true
-   * to NOT transfer the cache to the cloned instance.
-   */
-  clone(dontTransferCache?: boolean): this {
-    if (dontTransferCache) {
-      return this.type.getViewDU(this.node) as this;
-    } else {
-      const cache = this.cache;
-      this.clearCache();
-      return this.type.getViewDU(this.node, cache) as this;
-    }
-  }
 }
 
 export function isCompositeType(type: Type<unknown>): type is CompositeTypeAny {
