+++ packages/ssz/src/view/arrayComposite.ts
@@ -1,6 +1,7 @@
 import {getNodesAtDepth, Node, toGindexBitstring, Tree} from "@chainsafe/persistent-merkle-tree";
 import {ValueOf} from "../type/abstract";
-import {CompositeType, TreeView, CompositeView, CompositeViewDU} from "../type/composite";
+import {CompositeType, CompositeView, CompositeViewDU} from "../type/composite";
+import {TreeView} from "./abstract";
 
 /** Expected API of this View's type. This interface allows to break a recursive dependency between types and views */
 export type ArrayCompositeType<
