+++ packages/ssz/src/view/bitArray.ts
@@ -1,6 +1,7 @@
 import {Tree, Node} from "@chainsafe/persistent-merkle-tree";
 import {BitArray} from "../value/bitArray";
-import {CompositeType, TreeView} from "../type/composite";
+import {CompositeType} from "../type/composite";
+import {TreeView} from "./abstract";
 
 /**
  * Thin wrapper around BitArray to upstream changes to `tree` on every `this.set()`
